Added a different activity for scanning bluetooth (Still need to add a list of scanned devices) Added a floating button that starts activity and bluetooth scanning
