+++ app/src/main/java/com/teamhub/utd/hub/MainActivity.java
@@ -1,5 +1,6 @@
 package com.teamhub.utd.hub;
 
+import android.app.Activity;
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
 import android.content.BroadcastReceiver;
@@ -14,72 +15,24 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends Activity {
 
     // create variables
-    BluetoothAdapter bluetoothAdapter;
     Button button;
-
-    // Create a BroadcastReceiver for ACTION_FOUND.
-    private final BroadcastReceiver mReceiver = new BroadcastReceiver() {
-        public void onReceive(Context context, Intent intent) {
-            String action = intent.getAction();
-            if (BluetoothDevice.ACTION_FOUND.equals(action)) {
-                // Discovery has found a device. Get the BluetoothDevice
-                // object and its info from the Intent.
-                BluetoothDevice device = intent.getParcelableExtra(BluetoothDevice.EXTRA_DEVICE);
-                String deviceName = device.getName();
-                String deviceHardwareAddress = device.getAddress(); // MAC address
-                Log.i("Mac address:", deviceHardwareAddress + " " + deviceName);
-            }
-        }
-    };
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-
-        // Don't forget to unregister the ACTION_FOUND receiver.
-        unregisterReceiver(mReceiver);
-    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-
-        // set up bluetooth adapter
-        bluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
-
         // button to do something
         button = (Button)findViewById(R.id.button);
         button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                bluetoothOnOff();
+                Intent i = new Intent(MainActivity.this, UnPairedActivity.class);
+                startActivity(i);
             }
         });
-    }
-
-    private void bluetoothOnOff() {
-        if (bluetoothAdapter == null) {
-            Log.i("Error:", "Not working");
-        }
-        // if it is not enable then enable
-        if (!bluetoothAdapter.isEnabled()) {
-            Intent enableBtIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-            startActivity(enableBtIntent);
-
-            IntentFilter BTIntent = new IntentFilter(BluetoothAdapter.ACTION_STATE_CHANGED);
-            registerReceiver(mReceiver, BTIntent);
-        }
-        // if it is already enable then disable it
-        if(bluetoothAdapter.isEnabled()) {
-            bluetoothAdapter.disable();
-
-            IntentFilter BTIntent2 = new IntentFilter(BluetoothAdapter.ACTION_STATE_CHANGED);
-            registerReceiver(mReceiver, BTIntent2);
-        }
     }
 }
